proxy-file is written to slice dir

+++ scripts/c3grid/dummy-staging.sh
@@ -20,6 +20,14 @@ echo "$sfr" > $metafile
 #echo "NAY!" >&2
 #exit 1
 
+# example of proxy usage
+chmod 600 $X509_USER_PROXY
+globus-url-copy file://$(pwd)/$datafile gsiftp://mardschana2.zib.de/tmp/$datafile
+
+if [ "$?" -ne "0" ]; then
+    exit 1
+fi
+
 
 rm -f .GARBAGE
